a good chunk of the failed attempt at a lua version
decided that the strict ANSI was too great to overcome despite the
awesome syntax that could have been used to manage all the modules
