+++ bin/defib
@@ -1,57 +1,7 @@
 #!/usr/bin/env bash
-
-aed_pwd=`pwd -P`
-aed_default_base="${HOME}/.aed"
-aed_local_base="${aed_pwd}/.aed"
-aed_lua_dirs="#/lib/?.lua;#/lib/?/init.lua"
-aed_lua_cdirs="#/lib/?.so;#/lib/?/init.so" # TODO detect windows and use .dll
-
-export AED_BASE="${AED_BASE:=$aed_default_base}"
-export AED_LOCAL="${AED_LOCAL:=$aed_local_base}"
-export LUA_PATH="${aed_lua_dirs//#/$AED_BASE};${aed_lua_dirs//#/$AED_LOCAL}"
-export LUA_CPATH="${aed_lua_cdirs//#/$AED_BASE};${aed_lua_cdirs//#/$AED_LOCAL}"
-
-lua -l aed -e "main('${aed_pwd}')"
-
-#AED_LIB="${AED_BASE}/lib"
-#AED_MODULES="${AED_BASE}/modules"
-
-#AED_CONFIG="${HOME}/.aedrc"
-
-#AED_LOCAL_DIR="${PWD_DIR}/.aed"
-#AED_LOCAL_MODULES="${AED_LOCAL_DIR}/modules"
-#AED_FILE="${PWD_DIR}/Aedfile"
-
-#if [ ! -e "$AED_FILE" ]; then
-  #outp "You must create a project specific '$AED_FILE' file"
-  #exit 1
-#fi
-
-
-
-#echo "                 ______  _____"
-#echo "          /\    |  ____||  __ \\"
-#echo "         /  \   | |__   | |  | |"
-#echo "        / /\ \  |  __|  | |  | |"
-#echo "       / ____ \ | |____ | |__| |"
-#echo "      /_/    \_\|______||_____/"
-#echo
-#echo
-#echo "          Automated External "
-#echo "            Defibrillator "
-#echo
-
-#source "${AED_LIB}/utils.sh" || die "Could not find AED install"
-#debug "Sourcing core and platform from '${AED_LIB}'"
-#source "${AED_LIB}/core.sh"
-#source "${AED_LIB}/platform.sh"
-
-#outp "Shocking project to life with '${PKG_MGR}' on '${PLATFORM}' ... CLEAR!!!"
-
-#outp "Updating ${PKG_MGR}..."
-#update_pkg_mgr || die "Cannot to update package manager"
-
-#debug "loading ${AED_FILE}"
-#source "${AED_FILE}"
-
-#outp "It lives!!"
+set -e
+export AED_ROOT="${AED_ROOT:=${HOME}/.aed}"
+export LUA_BIN="${AED_ROOT}/lua/bin/lua"
+AED_CMD_NAME=$(basename "$0")
+AED_CMD_PATH=${AED_ROOT}/bin/${AED_CMD_NAME}.lua
+exec -a "$CMD_NAME" "$LUA_BIN" "$AED_CMD_PATH"
