+++ bootstrap.sh
@@ -1,0 +1,9 @@
+#!/usr/bin/env bash
+AED_ROOT=${AED_ROOT:=$HOME/.aed}
+
+if [ -d "${AED_ROOT}" ]; then
+  echo "Already installed"
+  exit 1
+fi
+
+
