+++ lib/bash/core.sh
