+++ lib/bash/platform.sh
