+++ lib/bash/platform/osx/homebrew.sh
