+++ lib/bash/utils.sh
