--- lib/utils.lua
@@ -1,3 +1,0 @@
-function fwrite (fmt, ...)
-  return io.write(string.format(fmt, ...))
-end
